fix: simplify functions and code
