+++ plugins/inputs/x509_cert/java_key_store.go
@@ -1,57 +1,13 @@
 package x509_cert
 
 import (
-	"bytes"
 	"crypto/x509"
-	"encoding/pem"
 	"fmt"
 	"os"
 
 	"github.com/pavlo-v-chernykh/keystore-go/v4"
 	"software.sslmate.com/src/go-pkcs12"
 )
-
-func isPEM(path string) bool {
-	content, err := os.ReadFile(path)
-	if err != nil {
-		return false
-	}
-
-	// Trim extra spaces before parsing
-	content = bytes.TrimSpace(content)
-
-	block, _ := pem.Decode(content)
-	return block != nil
-}
-
-func detectKeystoreFormat(path string) (string, error) {
-	file, err := os.Open(path)
-	if err != nil {
-		return "", fmt.Errorf("failed to open file: %w", err)
-	}
-	defer file.Close()
-
-	// Read first 4 bytes
-	magic := make([]byte, 4)
-	_, err = file.Read(magic)
-	if err != nil {
-		return "", fmt.Errorf("failed to read file magic bytes: %w", err)
-	}
-
-	// JKS magic number (big-endian): 0xFEEDFEED
-	if magic[0] == 0xFE && magic[1] == 0xED && magic[2] == 0xFE && magic[3] == 0xED {
-		return "jks", nil
-	}
-
-	// If not JKS, assume PKCS#12 (binary format)
-	// Since PKCS#12 does not have a magic number, we assume binary files are PKCS#12
-	if !isPEM(path) {
-		return "pkcs12", nil
-	}
-
-	// If the file is PEM, return "unknown" (so getCert() falls back to PEM logic)
-	return "unknown", nil
-}
 
 func (c *X509Cert) processPKCS12(certPath string) ([]*x509.Certificate, error) {
 	data, err := os.ReadFile(certPath)
@@ -64,8 +20,8 @@ func (c *X509Cert) processPKCS12(certPath string) ([]*x509.Certificate, error) {
 	if err != nil {
 		return nil, fmt.Errorf("failed to get password: %w", err)
 	}
-	defer password.Destroy()
 	passwordStr := password.String()
+	password.Destroy()
 
 	_, cert, caCerts, err := pkcs12.DecodeChain(data, passwordStr)
 	if err != nil {
