+++ plugins/inputs/x509_cert/sample.conf
@@ -7,7 +7,7 @@
             "/etc/ssl/certs/ssl-cert-snakeoil.pem",
             "/etc/mycerts/*.mydomain.org.pem", "file:///path/to/*.pem",
             "jks:///etc/mycerts/keystore.jks",
-            "jks:///etc/mycerts/keystore.p12"]
+            "pkcs12:///etc/mycerts/keystore.p12"]
 
   ## Timeout for SSL connection
   # timeout = "5s"
